lockfile.acquire doesn't accept a timeout of 0
Docs indicates that lock.acquire(timeout=0) would not block, but it
does (for linklockfile) because 0 is not truthy, and the line:

   timeout = timeout is not None and timeout or self.timeout

when timeout is 0 causes it to default to self.timeout, which by default is None.

So rewrite the condition better to take into account that timeout may
be set to 0.

Change-Id: Idf0d00977e79661e1eafd695d6e148e5f27e1840